Adds get_folds()
This handles an arbitrary number of class labels. The dataset is "sliced" by label and the corresponding observations with this label are divided up amongst the folds. Note that, in the worst case, a single fold may have L more training examples than the other folds, where L is the number of labels, since the modulo operation is used to assign an observation to a fold.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>