Adds create_all_split_tests() and create_split_tests()
Rather than storing the value of the feature in the ID3 decision tree, it is much more flexible to simply store the actual function that partitions the dataset to be passed to the left and right children of the node. In this way, continuous and discrete features can be treated the same way; the only difference is that continuous features result in functions evaluating based on inequality, while discrete test for equality.

The generation of the split tests needs to be modified since these tests are used to calculate the conditional entropy. For continuous variables, each partition value actually needs to generate both the less than or equal to and greater than tests, rather than just the less than or equal to test. The convention is that values of the dataset that satisfy the split test are passed to the left child of the node, while values that do not satisfy the split test are passed to the right child.

Signed-off-by: Ryan Tatton <rdt17@protonmail.com>